Added LoadRooms function
function that reloads the rooms from the files

+++ src/Game.cpp
@@ -1,17 +1,15 @@
 #include "..\include\Game.h"
 
 
-void Game::update(int act, int room)
+void Game::update(short act, short room)
 {
+	/*
     act_ = act;
     room_ = room;
 
-
-
     std::vector<sf::IntRect> t_new_room_tiles story.getNewRoomTiles();
 
-
-
+	*/
     return;
 }
